requirements.txt
pywin32>=306
PyQt5>=5.15.0
Pillow>=10.0.0
pyserial>=3.5
python-barcode>=0.15.1